// frontend/src/refinedApi/client/models/ComponentModel_Output.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
import type { PropertyModel_Output } from './PropertyModel_Output';
export type ComponentModel_Output = {
    name: string;
    input: Record<string, PropertyModel_Output>;
    output: Record<string, PropertyModel_Output>;
    description?: (string | null);
};